requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
apify_client==1.10.0
apify_shared==1.4.1
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.12.2
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
dateparser==1.2.1
demjson3==3.0.6
distro==1.9.0
facebook-scraper==0.2.59
fake-useragent==2.2.0
fastapi==0.104.1
feedparser==6.0.10
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.2
jiter==0.10.0
joblib==1.5.1
lxml==4.9.3
MarkupSafe==3.0.2
more-itertools==10.7.0
numpy==2.3.0
oauthlib==3.2.2
openai==1.81.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
playwright==1.52.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyee==11.1.1
pyparsing==3.2.3
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
regex==2024.11.6
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.0
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
stripe==12.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.15.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.24.0
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.22.0
